extracted BaseStream class
